v1.0.3 — CU7 (KB5096981) + Temmuz güvenlik bandı
Baseline CU6 -> CU7 (17.0.4065.4, 16 Temmuz 2026). 14 Temmuz GDR'lerinin
(KB5102333 / KB5101346) kapattığı yedi CVE kitaba işlendi.

Tüm kod örnekleri CU7 kurulu bir instance'ta yeniden koşuldu (92 script,
61 OK / 31 çevresel). Motor tarafından reddedilen ve düzeltilen maddeler:

Vector index:
- Salt-okunur kısıtı belgelendi (Msg 42231); ALLOW_STALE_VECTOR_INDEX CU7'de yok
- Clustered PK tek 4 baytlık INT olmalı (Msg 42217); chunk_id BIGINT -> INT
- FP16 sütun tipinde verilir: VECTOR(n, float16). PRECISION index seçeneği yok
- "En az 100 satır" şartı Azure SQL/Fabric'teki yeni index sürümüne ait

Change Event Streaming:
- sys.event_stream_groups / _group_tables / dm_event_stream_status yok
- sp_start_event_stream_group / sp_stop_event_stream_group yok
- İzleme sys.dm_change_feed_log_scan_sessions ve _errors üzerinden

Diğer:
- JSON_OBJECT/JSON_OBJECTAGG: VALUE degil iki nokta sözdizimi
- CAST('false' AS JSON) gecersiz; JSON kökte skaler kabul etmez
- REGEXP_MATCHES: match_position degil start_position
- sys.tables: is_ledger degil ledger_type
- sys.databases: is_change_tracking_on ve is_link_to_synapse_enabled yok
- sys.dm_db_vector_indexes ve dm_db_attestation_compute_capability CU7'de yok
- Resource Governor örnekleri idempotent hale getirildi

+++ kod-ornekleri/_ortak/00-demo-veritabani.sql
@@ -165,7 +165,9 @@ CREATE SCHEMA ai;
 GO
 
 CREATE TABLE ai.document_chunks (
-    chunk_id      BIGINT       NOT NULL IDENTITY PRIMARY KEY,
+    -- Vector index önkoşulu: clustered PK tek bir 4 baytlık INT sütun olmalı (Msg 42217).
+    -- BIGINT kullanılırsa CREATE VECTOR INDEX reddedilir.
+    chunk_id      INT          NOT NULL IDENTITY PRIMARY KEY,
     source_doc    NVARCHAR(200) NOT NULL,
     chunk_index   INT          NOT NULL,
     content       NVARCHAR(MAX) NOT NULL,
